Fix `read_dir` on a symlink to a directory ending in a `.`.
Rust's `Path` implicitly hides `.` components, however when `.` appears
as a component at the end of a path, it changes the meaning of the path.
Previously, we handled this by special-casing it in several places;
switch to prepending `CurDir` as a normal path component to the
components worklist so that it can be handled in a uniform manner.